Moved headers into agent folder at top leve for correct install of header files.
