+++ src/IAMQPWorkerSSL.cpp
@@ -1,7 +1,7 @@
-#include "IAMQPWorkerSSL.hpp"
-#include "IConnectionHandler.hpp"
-#include "IWorker.hpp"
-#include "SymbolMaps.hpp"
+#include "agent/IAMQPWorkerSSL.hpp"
+#include "agent/IConnectionHandler.hpp"
+#include "agent/IWorker.hpp"
+#include "agent/SymbolMaps.hpp"
 
 #include <string>
 #include <cstdint>
